implement: authentication :key:

+++ graphql/resolvers/booking.js
@@ -3,26 +3,35 @@ const Booking = require('../../models/booking');
 const { formatBooking, formatEvent } = require('./merge');
 
 module.exports = {
-  bookings: async () => {
+  bookings: async (args, req) => {
+    if (!req.isAuth) {
+      throw new Error('Unauthenticated!');
+    }
     const bookings = await Booking.find();
     return bookings.map(booking => {
       return formatBooking(booking);
     });
   },
-  bookEvent: async ({ eventID }) => {
-    const verifiyBooking = await Booking.findOne({ $and: [{ user: '5cb719721b51bb29c2c7bdbf' }, { event: eventID }] });
+  bookEvent: async ({ eventID }, req) => {
+    if (!req.isAuth) {
+      throw new Error('Unauthenticated!');
+    }
+    const verifiyBooking = await Booking.findOne({ $and: [{ user: req.userId }, { event: eventID }] });
     if (verifiyBooking) {
-      throw new Error('User has already booked for this event');
+      throw new Error('User has already booked for this event.');
     }
     const event = await Event.findById(eventID);
     const booking = new Booking({
-      user: '5cb719721b51bb29c2c7bdbf', // TODO: id from session
+      user: req.userId,
       event
     });
     const res = await booking.save();
     return formatBooking(res);
   },
-  cancelBooking: async ({ bookingID }) => {
+  cancelBooking: async ({ bookingID }, req) => {
+    if (!req.isAuth) {
+      throw new Error('Unauthenticated!');
+    }
     const booking = await Booking.findById(bookingID).populate('event');
     const event = formatEvent(booking.event);
     await Booking.findByIdAndDelete(bookingID);
